requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
asyncpg==0.29.0
pydantic==2.9.1
python-dotenv==1.0.1
sse-starlette==2.1.3
